Move files to features/walletconnect
