+++ src/features/walletconnect/__tests__/WalletConnectContext.test.tsx
@@ -5,13 +5,14 @@ import type { Web3WalletTypes } from '@walletconnect/web3wallet'
 import type { SessionTypes } from '@walletconnect/types'
 
 import { act, fireEvent, render, waitFor } from '@/tests/test-utils'
-import { WalletConnectContext, WalletConnectProvider } from '../WalletConnectContext'
-import WalletConnectWallet from '../WalletConnectWallet'
+import { WalletConnectContext } from '../WalletConnectContext'
+import WalletConnectWallet from '../services/WalletConnectWallet'
+import { WalletConnectProvider } from '../components/WalletConnectProvider'
 import { safeInfoSlice } from '@/store/safeInfoSlice'
 import { useAppDispatch } from '@/store'
 import * as useSafeWalletProvider from '@/services/safe-wallet-provider/useSafeWalletProvider'
 
-jest.mock('../WalletConnectWallet')
+jest.mock('../services/WalletConnectWallet')
 jest.mock('@/services/safe-wallet-provider/useSafeWalletProvider')
 
 jest.mock('@/hooks/safe-apps/useRemoteSafeApps', () => ({
