+++ src/features/walletconnect/components/WcHeaderWidget/index.tsx
