+++ src/features/walletconnect/components/WcLogoHeader/index.tsx
