+++ src/features/walletconnect/components/WcProposalForm/CompatibilityWarning.tsx
