+++ src/features/walletconnect/hooks/__tests__/useWalletConnectSearchParamUri.test.ts
