+++ src/features/walletconnect/services/utils.ts
@@ -1,6 +1,6 @@
 import type { ChainInfo } from '@safe-global/safe-gateway-typescript-sdk'
 import type { ProposalTypes, SessionTypes } from '@walletconnect/types'
-import { EIP155, BlockedBridges, WarnedBridges, WarnedBridgeNames } from '@/services/walletconnect/constants'
+import { EIP155, BlockedBridges, WarnedBridges, WarnedBridgeNames } from '@/features/walletconnect/constants'
 
 export const isPairingUri = (uri: string): boolean => {
   return uri.startsWith('wc:')
